Code refactoring, Optimization
