+++ frontend/src/App.jsx
@@ -2,12 +2,12 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 
-import Header from './components/Header';
-import Services from './components/Services';
-import Contact from './components/Contact';
-import Footer from './components/Footer';
-import Map from './components/Map'
-import About from './components/About';
+import Header from './components/header/Header';
+import Services from './components/services/Services';
+import Contact from './components/contact/Contact';
+import Footer from './components/footer/Footer';
+import Map from './components/map/Map'
+import About from './components/about/About';
 
 function App() {
   return (
