+++ frontend/src/components/carousel/CaroSlide.css
