+++ frontend/src/components/carousel/CaroSlide.jsx
