+++ frontend/src/components/hero/Hero.css
@@ -1,5 +1,5 @@
 .container__hero {
-  background-image: url("../../public/images/beanfarm.jpg");
+  background-image: url("../../../public/images/beanfarm.jpg");
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
