+++ frontend/src/components/hero/Hero.jsx
