+++ frontend/src/components/map/Map.css
@@ -1,5 +1,5 @@
 .container__map {
-	background-image: url('../../public/images/map.svg');
+	background-image: url('../../../public/images/map.svg');
 	background-repeat: no-repeat;
 	background-size: cover;
 	background-position: center;
