+++ frontend/src/components/nav/Navlist.jsx
@@ -1,0 +1,40 @@
+
+let id = 0
+export const navList = [
+	{
+		id: id++,
+		to: '../dashboard',
+		navName: 'Dashboard',
+		icon: false
+	},
+	{
+		id: id++,
+		to: '../agriculture',
+		navName: 'Agriculture',
+		icon: false
+	},
+	{
+		id: id++,
+		to: './#services',
+		navName: 'Services',
+		icon: false 
+	},
+	{
+		id: id++,
+		to: './#about',
+		navName: 'About us',
+		icon: false 
+	},
+	{
+		id: id++,
+		to: './#contact',
+		navName: 'Contact',
+		icon: false
+	},
+	{
+		id: id++,
+		to: '../login',
+		navName: 'LOGIN',
+		icon: true
+	},
+];
