+++ frontend/src/components/search/ProductsSearch.jsx
